requirement.txt
alembic==1.13.1
aniso8601==9.0.1
antlr4-python3-runtime==4.9.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
Deprecated==1.2.14
docker==7.1.0
efficientnet-pytorch==0.7.1
entrypoints==0.4
filelock==3.13.1
Flask==3.0.3
fonttools==4.53.0
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.43
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
huggingface-hub==0.23.4
hydra-core==1.3.2
idna==3.7
importlib_metadata==7.1.0
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.5
Mako==1.3.5
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.0
mkl==2021.4.0
mlflow==2.13.2
mpmath==1.3.0
munch==4.0.0
networkx==3.2.1
numpy==1.26.3
omegaconf==2.3.0
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
packaging==24.0
pandas==2.2.2
pillow==10.2.0
pretrainedmodels==0.7.4
protobuf==4.25.3
pyarrow==15.0.2
pycocotools==2.0.7
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
querystring-parser==1.2.4
requests==2.32.3
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
segmentation-models-pytorch==0.3.3
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.30
sqlparse==0.5.0
sympy==1.12
tbb==2021.11.0
threadpoolctl==3.5.0
timm==0.9.2
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.3.1+cpu
torchaudio==2.3.1+cpu
torchvision==0.18.1+cpu
tqdm==4.66.4
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
waitress==3.0.0
Werkzeug==3.0.3
wrapt==1.16.0
zipp==3.19.2
